Reduce warning messages for the test_noninteractive target
The changes were as follows:

* For the Fortran binding, drop declarations of some unused variables.

* For the Fortran examples, don't define an unused function.

* For the OCaml examples, drop some unneeded calculations.

* For the higher-level Octave binding that is implemented with Octave
  functions, replace calls to the deprecated (in Octave 4.4) toascii
  function with calls to the recommended replacement (the double
  function).  I assume this replacement is backwards compatible since
  I have discovered with web searchs that the double function has been
  documented back to octave-2.9 and likely even earlier.  However, I
  have no means to test octave versions less than 4.4.

  I speculate that toascii was deprecated in Octave 4.4 because its
  name was a horrible misnomer describing the reverse of what occurs
  which is conversion of each character in a string to the
  corresponding integer, but by default integers are internally stored
  in octave as double-precision floating point numbers so the actual
  conversion was from each character in the string to double (which I
  assume is why double is the recommended replacement).

As a result of these changes the only warning messages left that are
emitted by the test_noninteractive target are run-time "PLPLOT WARNING" messages
that have been present for a long time, and the following spurious warnings
emitted by gfortran:

1. argv_dynamic warning message.
/home/software/plplot/HEAD/plplot.git/examples/fortran/x01f.f90:51:0:

     character(len=:), dimension(:), allocatable :: argv_dynamic

Warning: ‘.argv_dynamic’ is used uninitialized in this function [-Wuninitialized]

I am not sure where the "." part of ".argv_dynamic" comes from in that
warning message.  In any case, I am virtually positive this warning is
spurious since argv_dynamic is an output argument of the routine
plget_arguments(argv_dynamic) which is the first mention of
argv_dynamic in an executable statement in this routine.

2. Spurious "PRIVATE" warning messages.
There are 56 of these due to a long-standing gfortran
bug (see
<https://gcc.gnu.org/bugzilla/show_bug.cgi?id=49111> and
<https://gcc.gnu.org/bugzilla/show_bug.cgi?id=64861>).  A typical
example is

/home/software/plplot/HEAD/plplot.git/bindings/fortran/plplot_small_modules.f90:108:8:

         function interface_strlen(s) bind(c, name='strlen')
        1
Warning: Symbol ‘interface_strlen’ at (1) is marked PRIVATE but has been given the binding label ‘strlen’

Tested by: Alan W. Irwin <airwin@users.sourceforge.net> on Linux
(Debian Buster) by configuring PLplot with the cmake options
-DBUILD_TEST=ON -DTRY_OCTAVE4=ON -DUSE_INCRTCL_VERSION_4=ON and
building the test_noninteractive target with octave 4.4.  All but the
"PLPLOT WARNING" and spurious fortran warning messages mentioned above
are now gone.  This overall test showed no obvious configure, build,
run-time, or PostScript difference report regressions as a result of
this commit.