Comprehensive test script: fix blank in fullpath bug
The dirname logic used to obtain SCRIPT_PATH needed to be modified
(using quotes) when the script was invoked using
a fullpath to the script that included a blank, e.g.,

"../plplot blank .git"/scripts/comprehensive_test.sh" ....

used to invoke the script in the test of the last commit.

Tested by: Alan. W. Irwin <airwin@users.sourceforge.net> on Linux
(Debian Buster) with the script invocation described above.