PLplot build system: bump minimum CMake version from 3.6.2 for all platforms to 3.7.2 for the Linux platform and 3.11.0 for non-Linux platforms
The motivation for these changes is to reduce build-system bugs by
forcing our users to use modern CMake (especially on non-Linux platforms where
older versions of CMake can be problematic).

With one exception, these minimum version choices mean that PLplot
users on modern Linux platforms and the major modern non-Linux
platforms should be able to rely on the official versions of CMake for
those platforms.  The exception to this PLplot user convenience is
Cygwin where the CMake packaging effort there has become moribund
(i.e., stuck on CMake 3.6.2 for much too long a time for what is
supposed to be a rolling release that keeps up with the latest free
software versions).  Because of this packaging problem for CMake on
Cygwin, this PLplot change means our Cygwin users will need to build
their own CMake with version = 3.11.0 or greater before they build
PLplot.

See also the commentary in the top-level CMakeLists.txt file for
details of the web sites checked for the CMake version that is
officially installed for various platforms.

I also took this opportunity to bump the minimum CMake version for
other minor projects within the PLplot source tree to 3.7.2 or 3.11.0
depending on whether for those projects I wanted to follow the
minimum version adopted for PLplot for either the Linux platform
or non-Linux platforms.