Build system: fix a bug in the use of the <TARGET_FILE:PLPLOT::pltcl> and $<TARGET_FILE:PLPLOT::plserver> generator expressions
The fix was to use these generator expressions only when ENABLE_tkX was true
since the shell script being configured only uses these locations for that
case.

Tested by: Alan W. Irwin <airwin@users.sourceforge.net> on Linux
(Debian Buster) by setting the DISPLAY environment variable to nothing
before configuring PLplot.  (An incorrect setting of DISPLAY in a
crontab task to submit a PLplot contract dashboard for CMake first
exposed these bugs).  For incorrect DISPLAY the Tk configuration of
PLplot fails (softly) with the result that ENABLE_tkX is false, which
means with this fix in place that the <TARGET_FILE:PLPLOT::pltcl> and
$<TARGET_FILE:PLPLOT::plserver> generator expressions are not used
which avoids the build-system error that occurred before this fix.