Build system: for Mac OS X drop user overide that imposes the dated -single_module linking option
Tested by Alan W. Irwin <airwin@users.sourceforge.net> on Linux
(Debian Buster) by configuring PLplot with cmake with no errors due to
this change.  However, that is just a minimal test (since this removed
override was a no-op on Linux) so the definitive test of this change
is requested on Mac OS X platforms.