Update scripts and instructions for generating local version of website
The scripts to help generate the website were updated with commentary
tweaks and an update to all make commands to use the
PARALLEL_BUILD_OPTION bash variable which was set to "-j10" at the
start of each script.  This value is suitable for my new hardware with
8 cores but could easily be modified by someone else with different
hardware whenever I step down as release manager.

I updated the instructions for the apache install and configuration
and instructions for the website generation for the release manager in
README.Release_Manager_Cookbook.  That file was also updated with all
references to "raven" (my old computer) replaced by "merlin" (my new
computer) and all references to Debian Jessie (the Debian version on
my old computer) replaced by Debian Buster (the Debian version on my
new computer).

Tested by: Alan W. Irwin <airwin@users.sourceforge.net> on Linux
(Debian Buster) by following the above apache install and
configuration instructions and the above website generation
instructions.  I did not do any of the detailed error checking that is
recommended for a release, but the result appeared to be a valid local
website that I could easily browse with the konqueror and firefox
browsers.