Move definitive source of the markdown version of our wiki pages from our SF wiki editing sessions to git control
This change provides automatic backup and (git) version
control of our wiki markdown source files, and it allows our developers to
use the editor of their choice to edit these files.

This change was implemented by inserting the following boilerplate header

-----------------8<-------------------
<!---
N.B.  The definitive Markdown source of this file is located in the
doc/wiki_source subdirectory of the PLplot source tree.  So only use
the ctrl-v and delete capabilities of the GUI file editor at
SourceForge to make changes to the SourceForge version of this file,
where ctrl-v is used to complete a cut and paste from the definitive
version of this file in the PLplot source tree that is being edited
with your favorite file editor, and delete used to remove extraneous
unmodified text whose changed form has been copied with the cut and
paste.
-->
-----------------8<-------------------

at the start of each of the SF markdown sources of our wiki pages
(accessible by editing each of our wiki pages at SourceForge) then
using that SF editor to select all text and use the legacy ctrl-c cut command for that
edit session to start a cut and paste to the equivalent file in the
doc/wiki_source subdirectory of our source tree.

Note that README.testing was an older version of Testing_PLplot
created in the same way.  So by copying the modern version of that
file to doc/wiki_source/Testing_PLplot by the above procedure and by
deleting README.testing we have effectively moved README.testing to
doc/wiki_source/Testing_PLplot and updated it to the latest SF
version.