Update from doxygen 1.8.8 to 1.8.13 in preparation for the release of 5.14.0
This commit is also part of the overall upgrade of our free software
dependencies to Debian Testing (currently the same as Debian Buster)
which is a rolling release whose free software versions are typically
close to the latest upstream releases.  (For example,
<http://www.doxygen.nl/download.html> states the latest doxygen
release right now is 1.8.14 so doxygen 1.8.13 that is packaged by
Debian Testing is behind that upstream version but not by very much.)

To upgrade our doxygen configuration to 1.8.13, I followed
the directions in (the updated) doc/README.doxygen and executed

doxygen -u Doxyfile.in

I also updated the testing directions in doc/README.doxygen.

Tested by: Alan W. Irwin <airwin@users.sourceforge.net> on Linux
(Debian Testing) by inspecting the Doxyfile.in change generated above
with "git diff" and following the additional test directions in the
updated doc/README.doxygen file.