Prep for 5.14.0: RELEASE_DATE=2018-12-12 and adjust for new version of plplot-doc-check
This update of RELEASE_DATE is consistent with the expected UTC date of the 5.14.0 release
and affects only the generated website documentation.

The new version (thanks to Hal Wang for this) of castxml-ready
plplot-doc-check code was required because the old version of that
code used the gccxml app which is not available for Debian Testing
because it has been completely superseded by castxml.  So this change
is part of the overall project to make PLplot consistent with the
modern versions of free software.

Tested by: Alan W. Irwin <airwin@users.sourceforge.net> on Linux
(Debian Testing) by building a local version of the website using

time (echo irwin$'\n'irwin$'\n'merlin$'\n'/home/irwin/public_html/plplot$'\n'yes |scripts/generate_website.sh)

and checking that the generated documentation had the correct release date.

In addition, PLplot was configured with the cmake option
-DADD_SPECIAL_CONSISTENCY_CHECKING=ON which configures the
check_api_xml_consistency target.  Building that target runs
the castxml-ready plplot-doc-check appropriately.  Building
that target with an artificially introduced local api.xml
inconsistency quickly identified that inconsistency.  So
this golang app seems to be a good one for finding such inconsistencies.
Finally, building that target without an introduced inconsistency
showed api.xml is consistent with plplot.h which is a good thing
to establish since it means our DocBook-generated documentation
concerning our API is consistent with our actual API!