Drop some additional symbols when generating comparison for bindings/fortran/plplotfortran_ifort.def
The extra line in the sed script is

/^.*MOD___/ d

which deletes every line with that pattern in it.  As a result the
following raw symbols (directly from the nm --defined-only command)
are dropped

000000000004B840 T __PLPLOT_GRAPHICS_MOD___COPY_PLPLOT_GRAPHICS_PLGRAPHICSIN
000000000004F9A0 R __PLPLOT_GRAPHICS_MOD___DEF_INIT_PLPLOT_GRAPHICS_PLGRAPHICSIN
0000000000057C80 D __PLPLOT_GRAPHICS_MOD___VTAB_PLPLOT_GRAPHICS_PLGRAPHICSIN
000000000004B820 T __PLPLOT_TYPES_MOD___COPY_PLPLOT_TYPES_PLCGRID
000000000004B810 T __PLPLOT_TYPES_MOD___COPY_PLPLOT_TYPES_PLFGRID
000000000004FA20 R __PLPLOT_TYPES_MOD___DEF_INIT_PLPLOT_TYPES_PLCGRID
000000000004F9F0 R __PLPLOT_TYPES_MOD___DEF_INIT_PLPLOT_TYPES_PLFGRID
0000000000057D00 D __PLPLOT_TYPES_MOD___VTAB_PLPLOT_TYPES_PLCGRID
0000000000057CC0 D __PLPLOT_TYPES_MOD___VTAB_PLPLOT_TYPES_PLFGRID

I could modify the sed script to process these symbols, but Arjen
seems content with the current
bindings/fortran/plplotfortran_ifort.def file that does not have the
processed version of these files.  So this one-line sed script change
means the generated comparison for that file is identical with the
original with the result that building the
check_ifort_definitions_file target succeeds.

Tested by: Alan W. Irwin <airwin@users.sourceforge.net> on Linux
(Debian Testing) by building the check_ifort_definitions_file target
with success.