Limited adjustments to the page on testing PLplot.
The edits consist mostly of splitting up the long paragraphs into smaller ones. Also adding an example on how to use the comprehensive testing script in the most minimal fashion. This prepares for a further reorganisation of the page.

Tested by Arjen Markus, via visual inspection of the wiki page.