Allow all qt devices to be used from a qt application
All qt devices need a qApp in order to work and the purpose of
initQtApp (which is called from all plD_init* routines *other than
plD_init_extqt* that are implemented in drivers/qt.cpp) is to create a
qApp (and the associated argv memory it needs) if a qApp doesn't exist
already (either created by a previous call to initQtApp or created
independently in an external qt application).  And the purpose of
closeQtApp (which is called from all the plD_tidy_* routines
implemented in drivers/qt.cpp) is to delete the qApp and associated
(argv) memory if this is the last use of the qApp that was initially
created in initQtApp, but not delete qApp if it was created
independently.  The static variable appCounter (protected by a mutex
to make it thread-safe if qt devices are being used in multiple threads)
is used to keep track of last use where appCounter will be zero.

This logic should work fine for non-qt applications that simply rely
on the PLplot C library to provide the qt devices (dynamically or as
part of the library) since for the last use case appCounter is zero
and therefore the internally created qApp and associated argv are
correctly deleted.  And it also works fine for qt applications (those
which depend on libplplotqt and which must necessarily create their
own qApp) if those applications only use the extqt device because in
that case initQtApp is never called so that when extqt is tidied up
appCounter ends up as -1 so the qApp created independently by that qt
application is (rightly) not deleted.  That leaves the problem of a qt
application that use non-extqt qt devices where prior to the present
change appCounter ends up as zero on last-use case, and the
independent qApp created by that qt application gets incorrectly
deleted.  To fix that case, this commit simply adds 1 to appCounter
in initQtApp when on first use case an external qApp is detected.

Tested by: António R. Tomé <antoniortome@users.sourceforge.net> on Linux
(openSUSE leap 15.0) by building an running his own qt application that
uses qt devices other than extqt without issues.