Clean up logic for creation and destruction of internal qApp
The previous implementation of this logic was being made unnecessarily
complicated and non-robust by not distinguishing clearly between the
case where initQtApp creates an internal qApp and the case where the
qApp is created externally.  Because of this issue, appCounter counted
users of qApp regardless of whether that qApp had been created
internally or externally; in one case (qt application using the extqt
device) we were relying on that counter being negative in closeQtApp
(the routine used to delete qApp if that was created internally by
initQtApp) in order to not destroy the external qApp always used in
that case, and in another case (qt application using non-extqt qt
devices) we were adding 1 to appCounter to avoid incorrect destruction
of the external qApp by closeQtApp.  Also as a result of this issue,
the case of a qt application using both extqt and a non-exqt qt device
would not work (since appCounter then ends up as zero so that the
external qApp used for this case gets incorrectly destroyed).

So for this cleanup I keep track of whether qApp is externally created
or internally created with the bool internal_qApp global variable,
rename appCounter to internal_qAppCounter, and consistent with that
name change make sure that variable
is only changed if internal_qApp is true.  This change allows us to
not have special conditions on the initial value of what is now called
internal_qAppCounter, and if an internal qApp is created, the last-use
case is unequivocally signalled by internal_qAppCounter == 0 *so long
as for each qt device that happens to use an internal qApp each initQtApp call in plD_init_* is
matched by a corresponding call to closeQtApp in plD_tidy_*.

That condition was previously met. And since extqt is the one qt
device that never uses an internal qApp, calls to initQtApp and
closeQtApp for that device would effectively become no-ops with the
new more robust version of the code.  But to keep the code as simple as
possible we continue as before not calling initQtApp from
plD_init_extqt, but we change from calling closeQtApp in
plD_tidy_extqt to dropping that call.

Tested by: Alan W. Irwin <airwin@users.sourceforge.net> on Linux
(Debian testing) by building the test_noninteractive, test_interactive,
and test_qt_example targets without configure, build, run-time, or
difference report issues. (The test_qt_example target is currently
not run as part of test_interactive because of occasional segfault issues
on exit from qt_example, but for this test we did not encounter
one of those, and qt_example appeared to work well despite the
known memory management issues it has.)

N.B. Our current suite of tests does not include the case of a qt
application that uses non-extqt qt devices.  However, António R. Tomé
has developed such applications so we request that he tests the
current commit with those applications.