Exported two functions from plstream.cc into the plplotcxx lib/dll for dll builds
The functions Contourable_Data_evaluator() and Coord_Xform_evaluator()
were not prefixed by PLDLLIMPEXP_CXX. This means that for dll builds of
the library they were not available to the users (although code compiled
it would not link). The Coord_Xform_evaluator() function is for use at
least in the plstream::vect method for mapping between the Coord_Xformer
class and the void * input required for the API.
Contourable_Data_evaluator() similarly mapps between the Contourable_Data
class and void * for ingestion into the API.

Both the functions should be available to the user, hence the need for
the export.