Update instructions for wiki changes
The previous instructions were for the case where changes were
made in the wiki markdown source in the local git working directory
and propagated to the appropriate SF wiki edit session.

I have now documented a second case where changes are made to the
appropriate SF wiki edit session of the markdown wiki source
and propagated to the local git working directory.