Drop trailing blank from wiki source for Configure_PLplot_for_the_Visual_Studio_IDE
This issue was discovered and fixed by running
scripts/remove_trailing_whitespace.sh.

Note this change to the git version of
Configure_PLplot_for_the_Visual_Studio_IDE was accompanied by the same
one-character change to the (edited) SF version to keep them
identical.  Furthermore, the tag on the SF version was changed from
2014 to 2019 to reflect the year when the page was last extensively
edited (by Phil in this case).