Remove unneeded libplplot dependency on qhull headers
Before this commit, libplplot (via src/plgridd.c) implemented a
run-time check concerning the size of the realT type used by libqhull
which forced this library be dependent on qhull headers even though
(unlike the csironn library) libplplot is not directly dependent on
the qhull library.  This commit removes this unnecessary complication
by implementing a CMake-based configuration check concerning the size
of the realT type used by libqhull.  If that check is not satisfied,
then our build system now simply drops use of qhull by dropping the
build and use of the csironn library.

N.B. src/plgridd.c now does include source code that depends directly
on both the qhull headers and qhull library.  However, this code is
protected by by the conditional compilation macro PLPLOT_NONN.  This
macro (previously named NONN) has not been #defined in our official
source tree for more than a decade which is why our build system
currently ignores the PLPLOT_NONN possibility and simply directly
links libplplot to the csironn library instead of using this internal
functionality.  Of course, if someone wants to try out use of this
internal functionality, the required build system changes would nearly
be trivial.  That is, the build system would need changes to provide
the PLPLOT_NONN cmake option, convert it to a C macro of the same name
which is #defined or not depending on whether the option is true or
not, and if the option is true drop the build of the csironn library
and drop the libplplot dependency on that library.

This commit also provides a change (preferring the libqhull form of
includes over the old qhull form) so that the headers for a local
build of upstream qhull (which always uses the libqhull form now)
can be found consistently with the library for
that local build that is found.

Tested by: Alan W. Irwin <airwin@users.sourceforge.net> on Linux
(Debian Testing) using the same extensive tests described for
the prior commit which used all these qhull-related changes
to PLplot that were in the working directory but not committed
at that time.