Build system: minor cleanup of the cgm device driver configuration
I recently discovered that the cgm_TARGETS CMake variable was not used
anywhere in the build system so there is no need to set it.
Therefore, since all that is done by cmake/modules/cgm.cmake is to set
that unneeded variable I have dropped this file from the cgm device
driver configuration.