Summarize my (Alan W. Irwin's) recent Linux (Debian Testing) comprehensive test result in our wiki (part II)
Part I of this change was a previous commit (with the same core title
as this commit) which summarized my (good) comprehensive test results
for CMake-3.13.2 (our minimum allowed CMake version) in our
"Testing_Reports" wiki page.  This commit (Part II) does the same
thing for my (good) comprehensive test results for CMake-3.14.4 (the
latest release of CMake that is available at this time).  So these two
commits document in our wiki that the current development version of
PLplot works well on Linux (Debian Testing) for the range of CMake
versions from 3.13.2 to 3.14.4.