Added instructions for unsupported paths to wiki
It was found that the build system cannot cope with paths containing a
space followed by a hyphen " -". Fixing this problem is likely to be
exceedingly time consuming for little benefit. so it was decided that
we would make every effort to warn users of the problem instead. This
change constitutes part of that warning.