Removed plexit call associated with not finding the .fnt files
To date, if plplot could not find the .fnt files, plexit was called
causing the whole program to exit. This commit implements warnings
instead and the behaviour is as follows if the .fnt files are not found
at initialisation:
A warning is generated stating the file was not found
For drivers which render their own text and provide their own string size
information, no other warning will be seen and renderign will proceed as
usual
For drivers which render their own text but do not provide their own string
size information, a warning will be seen when PLplot tries to determine the
string size using the hershey fonts. Text centring may be effected, however,
some drivers that report not being able to provide string size information
do their own centring, so in this case their is no impact on rendering.
For drivers which do not render their own text, no text will be rendered and
warnings will be issued.

If the .fnt files are found at initialisation then obviously everything
proceeds as normal.

This change was tested with the wxWidgets driver (renders its own text and
provides string size information), the ps driver (renders its own text but
does not provide string size information) and the wingcc driver (does not
render its own text). All three drivers were tested with .fnt files available
and with .fnt files not available. All three drivers produced the expected
results listed above. The tests were performed on a Windows machine.