D language support topic: Add D language support files from the cmake-d project
These files were downloaded from the [cmake-d project](https://github.com/dcarp/cmake-d)
as of 2019-08-14.

This commit is the first in the D languate support topic branch
devoted to replacing our ancient fork of the cmaked project for D
language support (which currently works for gdc, but which is
completely broken for the dmd compiler) by a fork of the much more
modern cmake-d project which will cleanly support the gdc, ldc, and
dmd compilers.