nistcd: Move from include_directories to target_include_directories for all nistcd targets
CMake best practices recommend all include_directories commands be replaced by
corresponding target_include_directories commands.  This commit implements that
change for the nistcd library and executables.

For libnistcd the PUBLIC header location is defined appropriately for
both the build-tree and install-tree cases and there are no PRIVATE
header locations.  The executables (which are all just built in the
build tree and not installed) all link to libnistcd so use its PUBLIC
header location to find the required nistcd header locations.  That
is, a call to target_include_directories is not needed for any of the
nistcd executables.  Note also because of the PUBLIC header location
for libnistcd that no special provision to find nistcd headers is
required by anything that links to libnistcd such as the PLplot cgm
device driver or libplot itself (when nondynamic device drivers are
used).

Tested by: Alan W. Irwin <airwin@users.sourceforge.net> on Linux
(Debian Buster=Stable) by configuring PLplot using -DPLD_cgm=ON
(which makes all nistcd targets available for building) and
building the test_nistcd target.  That built the nistcd
library and associated executables without issues (and
also successfully run-time tested all those executables).