Build system: Only update DRIVERS_LINK_FLAGS for the wxwidgets case if -DPLD_wxwidgets=ON
This commit fixes a bug in our build-system logic for the wxwidgets
case where it was possible to update DRIVERS_LINK_FLAGS with wxwidgets
information for -DENABLE_wxwidgets=OFF (which always forces
-DPLD_wxwidgets=OFF).  As a result of this fix our linking flags for
the STATIC libplplot case do not get contaminated with wxwidgets
information when libplplot does not include the wxwidgets device
driver.

Tested by: Alan W. Irwin <airwin@users.sourceforge.net> on Linux
(Debian Buster=Stable) by configuring PLplot with
-DBUILD_SHARED_LIBS=OFF -DENABLE_wxwidgets=OFF which is the
configuration that exposed this bug before, and VERBOSE=1 building the
plplot target.  The result showed no contamination of that build by
wxwidgets linking flags which is the desired result we did not have
before this commit.