test_d: Update README for this small self-contained project to test D language support
Explain the current status of upstream dmd which is testing two
methods (-review-noXlinker and -Xcc) of allowing dmd users to
communicate with the C compiler that dmd uses for linking before
deciding on which of those two methods will be available for the next
dmd release.  Also explain how to get access to two different dmd
versions that implement these two different methods.

Note my further plan is to create a CMake option for test_d that gives
users of it a chance to test the -pthread option (in raw form for gdc,
and in -Xcc-pthread form for both ldc2 and this version of dmd).  But
that is not available for now so all comprehensive testing of test_d
currently does not check -pthread, i.e., it simply checks that the C
compiler and various D compilers work correctly together to build (for
the shared libraries and static libraries for both build-tree and
install-tree conditions) a minimal project consisting of a C-language
library, a D-language library that links to that C library, and a
D-language executable that links to that D-language library.

Tested by: Alan W. Irwin <airwin@users.sourceforge.net> on Linux (Debian Buster)
by invoking the comprehensive test script for this project in the 3 ways documented
by README (where the dmd version used corresponded to the -Xcc version).  All three
comprehensive tests were a success including the one for the -Xcc version of dmd.
Note this comprehensive test does not (yet, see above) test the -pthread case
so this test did not test the -pthread option for gdc or the -Xcc-pthread option for
dmd or ldc2, but such a test is in the planning stages (see above).