test_d: change the C library to use the -pthread link option and test that case for all 3 D compilers
The change to the C library is designed to expose the issue of static
D linking with the -pthread linking option for a mixture of C- and
D-compiled code where the linking is done by either dmd or ldc2.
(There is no such linking issue with gdc because that D compiler
understands the -pthread option.)

A nice CMake-based fix for when this issue is exposed would ideally
use the conditional linking generator expressions that are part of
[this CMake feature
request](https://gitlab.kitware.com/cmake/cmake/issues/19757).  But
since those CMake generator expressions are not implemented yet, I
instead implemented a proof-of-concept workaround for conditions which
expose this issue.  This workaround requires building the C library
identically twice with two different names (one with a suffix of
"_static_D_linking") with the only difference being that the
INTERFACE_LINK_LIBRARIES property of the suffixed version is modified
(when the D compiler is not gdc) to replace the link to
Threads::Threads (if that is equivalent to -pthread) by the linker
option -Xcc=-pthread (which the dmd and ldc2 D compilers transform to
the necessary -pthread option for the C compiler they both use for
linking).  This workaround of building two versions of the same
library is obviously ugly so I plan to shift to using the generator
expressions if/when the CMake developers implement those.  However, as
shown below, this workaround works for now.

Tested by: Alan W. Irwin <airwin@users.sourceforge.net> on Linux
(Debian Buster=Stable) by invoking the comprehensive test script and
evaluating the results of that script for this test_d project for each
of dmd, ldc2, and gcd in the way documented by README (where the
chosen special version of dmd has the needed -Xcc capability).  For
each of the three D compilers the results were perfect.

These perfect comprehensive test results show that dmd and ldc2
(because of the above workaround) and gdc (which does not need the
above workaround) all work perfectly when -pthread is a linker option
for the C library.  So these are quite promising results, and the next
step is to implement the same workaround for the plplot project case
to solve the same issue that has now been exposed for test_d for the
dmd and ldc2 D compilers.