test_d: Improve implementation of how the -pthread option is handled for statically linked D examples
The fundamental improvement here is to use a generator expression

$<IF:$<COMPILE_LANG_AND_ID:D,DigitalMars,LDC>,-Xcc=-pthread,-pthread>

to help choose the correct transformation (-Xcc=-pthread for dmd and
ldc2 and -pthread for gdc) of the -pthread link option for the various
D compilers that are used for static linking of C- and D-compiled
objects with a D compiler.  This implementation requires two
alternatives to the normal C library which are the "alt_threads" and
"alt_interface" forms of that library.  The "alt_threads" form is a
copy of the normal C library except that all elements of
INTERFACE_LINK_LIBRARIES for that full library are the same as the
equivalent property for the normal C library with all equivalents of
the -pthread link option removed.  Thus, this library is used simply
to provide C objects for static linking with the D compiler without
propagating the untransformed -pthread linking option in any form.
The "alt_interface" form is an interface library that has a single
INTERFACE_LINK_OPTIONS property which conveniently provides the
generator expression described above whereever it is needed.

N.B. If/when [CMake bug
19757](https://gitlab.kitware.com/cmake/cmake/issues/19757) is
addressed these "alt_threads" and "alt_interface" variants of the C
library will no longer be needed, and we should be able to use the
generator expression directly in the INTERFACE_LINK_LIBRARIES property
of the normal C library to achieve a much more elegant solution to the
"-pthread issue" for static linking of C- and D-compiled objects with
a D compiler.

In any case, this "two alt C library" approach is working (see below)
for the test_d project and is also the approach I plan to use for the
more complex plplot project case for both D and Fortran -- see the
discussion in the above CMake bug report concerning the similar Fortran
issue with -pthread.

I also took this opportunity to update the test_d README
file to reflect the new master branch location of the -Xcc=
dmd option and to better describe my dmd build procedure.

Tested by: Alan W. Irwin <airwin@users.sourceforge.net> on Linux
(Debian Buster=Stable) by invoking the test_d comprehensive test
script and evaluating the results of that script for this test_d
project for each of dmd, ldc2, and gcd in the way documented by README
(where the chosen special version of dmd has the needed -Xcc
capability).  Th evaluation of those three comprehensive tests using
the documented procedure (which includes checking the correct -pthread
transformation has been done for each D compiler) in the README showed
no issues.  So the present improvement has introduced no test_d
regressions on Linux, and test_d should continue to be ready for
testing on other platforms.