Add CMake-based build system for ndiff utility
This build system was copied from my work (along with Andrew Ross)
in cmake/epa_build/ndiff.  But in addition I have updated that
build system to modern cmake.  This
is currently a standalone build system, and I leave it as a future
exercise to see how to optionally make it part of the PLplot
build system as well so that PLplot's make, make install, and
make test will include ndiff.  It's possible that is as simple
as adding "add_subdirectory(ndiff)" to utils/CMakeLists.txt,
but I leave it as a future exercise to find out the best way
to do this if that simple change is not sufficient.

Tested by: Alan W. Irwin <airwin@users.sourceforge.net> on Linux
(Debian Buster = Stable) by following the directions in
utils/ndiff/README.CMake_build_system.  All went well with
configuration, build, test, and install.  And I subsequently used the
installed version of ndiff successfully to evaluate some (slightly)
changed results for my FreeEOS project.