Update memqt_example
Some fragmentary (and therefore non-working) support for test_xor mode was
included in error in the original version of this example, and this
commit fixes that issue.  In addition, I have moved from the
semitransparent almost black background that was used before
to the standard opaque black because semitransparent backgrounds
are not rendered correctly on my Debian Buster platform, and opaque
black is what is used for the standard example 1 code that is
the source of much of the logic in this memqt_example.

Tested by: Alan W. Irwin <airwin@users.sourceforge.net> on Linux
(Debian Buster = Stable) by configuring PLplot and building the
test_memqt_example target without configuration, build, or run-time
issues with the resulting GUI display of the memqt and pngqt results
looking identical with each other and also the results of standard
example 1.