memqt device: Always make sure the memory for this device is initialized and once again enable plscolbg(a) for this device
It is important to remember that the memory area for memqt must always
be initialized with either a background image or an RGBA colour or
otherwise there are memory management issues.  Furthermore, when
plscolbg(a) is enabled for a device (as now is the case for the memqt
device), it is important to remember that PLplot *always* applies an
RGBA background (which defaults to opaque black).  But PLplot also
allows full control of those background RGBA colours.  So this commit
modifies memqt_example to be consistent with those two constraints
while not interfering with functionality of the example in the
slightest.  It does this by making full use of absolutely transparent
raw buffer (or Qt5 image) initialization when required and absolutely
transparent PLplot background when required.

For this commit I also modified scripts/style_source.sh to apply
to all source code in examples/c++/memqtTest and used that
script to style that source code.

Tested by: Alan W. Irwin <airwin@users.sourceforge.net> on Linux
(Debian Buster = Stable) by configuring PLplot and building the
test_memqt_example target without configuration, build, or run-time
issues.  For that run-time test I tested every action available from
the GUI without issues.  Furthemore, I ran

valgrind --num-callers=40 examples/c++/memqt_example ../plplot.git/examples/c++/memqtTest/aurora.png

six different times where in each case I executed just one of the actions, and in all cases
valgrind had a report similar to the following:

==32100== Memcheck, a memory error detector
==32100== Copyright (C) 2002-2017, and GNU GPL'd, by Julian Seward et al.
==32100== Using Valgrind-3.14.0 and LibVEX; rerun with -h for copyright info
==32100== Command: examples/c++/memqt_example ../plplot.git/examples/c++/memqtTest/aurora.png
==32100==
==32100==
==32100== HEAP SUMMARY:
==32100==     in use at exit: 395,593 bytes in 6,636 blocks
==32100==   total heap usage: 207,141 allocs, 200,505 frees, 41,529,600 bytes allocated
==32100==
==32100== LEAK SUMMARY:
==32100==    definitely lost: 6,800 bytes in 29 blocks
==32100==    indirectly lost: 2,205 bytes in 103 blocks
==32100==      possibly lost: 352 bytes in 1 blocks
==32100==    still reachable: 386,236 bytes in 6,503 blocks
==32100==         suppressed: 0 bytes in 0 blocks
==32100== Rerun with --leak-check=full to see details of leaked memory
==32100==
==32100== For counts of detected and suppressed errors, rerun with: -v
==32100== ERROR SUMMARY: 0 errors from 0 contexts (suppressed: 2 from 2)

So these reports mean there are no serious remaining memory management
errors for the memqt device and memqt_example.  Furthemore, I ascribe
the memory leak issues reported by valgrind to Qt5 itself rather than
our own code so there appears to be nothing else for PLplot to do with
regard to memory managment for the memqt device or memqt_example.

N.B. If I did not initialize the raw memqt buffer, this report was
*much* longer with many management errors reported.  So that
initialization is absolutely essential, but it turns out it only needs
to be done for just one action because for all the other actions the
memqt buffer is initialized by either aurora.png or a Qt5 transparent
image.